diversas melhorias após integração com a Fermentec

+++ management/commands/bluedata_cmd.py
@@ -1,10 +1,8 @@
 import os
-import tempfile
+import fileinput
 
 from django.core.management.base import BaseCommand
-
 from shutil import copyfile
-import fileinput
 
 
 class Command(BaseCommand):
@@ -20,74 +18,68 @@ class Command(BaseCommand):
         app_path = options['app_path']
         model_name = options['model_name']
         base_url = options['base_url']
-
         self.stdout.write('Starting process ...')
 
         try:
             self.stdout.write('app_path: {}'.format(app_path))
             self.stdout.write('model_name: {}'.format(model_name))
             self.stdout.write('base_url: {}'.format(base_url))
+            self.stdout.write('xxx')
 
             # admin.py
-            src = 'bluedata/templates/bluedata_modelos/admin.py'
-            dst = '{}/admin.py'.format(app_path)
-            copyfile(src, dst)
-            replace_infile(dst, '[app_path]', app_path)
-            replace_infile(dst, '[model_name]', model_name)
-            delete_first(dst)
-            self.stdout.write('>>> {} criado com sucesso'.format(dst))
+            create_file(self, 'admin', app_path, model_name, base_url)
 
             # filters.py
-            src = 'bluedata/templates/bluedata_modelos/filters.py'
-            dst = '{}/filters.py'.format(app_path)
-            copyfile(src, dst)
-            replace_infile(dst, '[app_path]', app_path)
-            replace_infile(dst, '[model_name]', model_name)
-            delete_first(dst)
-            self.stdout.write('>>> {} criado com sucesso'.format(dst))
+            create_file(self, 'filters', app_path, model_name, base_url)
 
             # forms.py
-            src = 'bluedata/templates/bluedata_modelos/forms.py'
-            dst = '{}/forms.py'.format(app_path)
-            copyfile(src, dst)
-            replace_infile(dst, '[app_path]', app_path)
-            replace_infile(dst, '[model_name]', model_name)
-            delete_first(dst)
-            self.stdout.write('>>> {} criado com sucesso'.format(dst))
+            create_file(self, 'forms', app_path, model_name, base_url)
 
             # models.py
-            src = 'bluedata/templates/bluedata_modelos/models.py'
-            dst = '{}/models.py'.format(app_path)
-            copyfile(src, dst)
-            replace_infile(dst, '[app_path]', app_path)
-            replace_infile(dst, '[model_name]', model_name)
-            delete_first(dst)
-            self.stdout.write('>>> {} criado com sucesso'.format(dst))
+            create_file(self, 'models', app_path, model_name, base_url)
 
             # urls.py
-            src = 'bluedata/templates/bluedata_modelos/urls.py'
-            dst = '{}/urls.py'.format(app_path)
-            copyfile(src, dst)
-            replace_infile(dst, '[app_path]', app_path)
-            replace_infile(dst, '[model_name]', model_name)
-            replace_infile(dst, '[base_url]', base_url)
-            delete_first(dst)
-            self.stdout.write('>>> {} criado com sucesso'.format(dst))
+            create_file(self, 'urls', app_path, model_name, base_url)
 
             # views.py
-            src = 'bluedata/templates/bluedata_modelos/views.py'
-            dst = '{}/views.py'.format(app_path)
-            copyfile(src, dst)
-            replace_infile(dst, '[app_path]', app_path)
-            replace_infile(dst, '[model_name]', model_name)
-            replace_infile(dst, '[base_url]', base_url)
-            delete_first(dst)
-            self.stdout.write('>>> {} criado com sucesso'.format(dst))
+            create_file(self, 'views', app_path, model_name, base_url)
 
         except Exception as ex:
             self.stdout.write('Erro ocorrido: {}'.format(ex))
 
         self.stdout.write('Ending process ...')
+
+
+def create_file(self, file, app_path, model_name, base_url):
+    try:
+        src = 'bluedata/templates/bluedata_modelos/{}.py'.format(file)
+        dst = '{}/{}.py'.format(app_path, file)
+        bkp = '{}/_{}.py'.format(app_path, file)
+
+        # backup do arquivo original
+        if os.path.isfile(dst):
+            copyfile(dst, bkp)
+            self.stdout.write('>>> backup: {}'.format(bkp))
+
+        # substitui por modelo
+        if os.path.isfile(src):
+            copyfile(src, dst)
+            self.stdout.write('>>> overwrite: {}'.format(dst))
+
+        # altera as chaves nos arquivos
+        replace_infile(dst, '[app_path]', app_path.replace('/', '.'))
+        replace_infile(dst, '[model_name]', model_name)
+        replace_infile(dst, '[base_url]', base_url)
+
+        # # remove a linha de comentário
+        delete_first(dst)
+
+        # sucesso
+        self.stdout.write('>>> {} criado com sucesso'.format(dst))
+        self.stdout.write('xxx')
+
+    except Exception as e:
+        print(e)
 
 
 def replace_infile(dst, text_to_search, replacement_text):
